[graphite] Mark BindBufferInfo bool operator explicit; make RenderStateFlags uint8_t
Ran into surprising compilation "successes" in the follow up CL because
the non-explit bool operator was allowing BindBufferInfo to implicitly
convert all the way to a uint32_t...

Also adds a more constrained backing type to RenderStateFlags so that
it can pack with other byte fields in DrawWriter.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1085356
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ src/gpu/graphite/DrawTypes.h
@@ -188,7 +188,7 @@ enum class DstUsage : uint8_t {
 };
 SK_MAKE_BITMASK_OPS(DstUsage)
 
-enum class RenderStateFlags : unsigned {
+enum class RenderStateFlags : uint8_t {
     kNone                   = 0b0000,
     kFixed                  = 0b0001,   // Uses explicit DrawWriter::draw functions
     kAppendVertices         = 0b0010,   // Appends vertices
